Allow transformations at record (de)serialization time
A new PailStructure wrapping an exisiting instance is provided.
The TransparentConversionPailStructure behaves like the wrapped structure
but applies transformations to (de)serialization output.
In order for Pail to ignore this transformation, the structure field of
a pail metadata file is set to the name of the underlying structure.
This is achieved by replacing usages of "getClass.getString" by a new
method in PailStructure.
To be less disruptive, a default behaviour mimicking the current
logic is provided. It is achieved by turning PailStructure into an
abstract class: API BREAK.